move hittable_list generation out from simple tracer's main cpp
