+++ src/common/hittable_list.cpp
@@ -1,0 +1,21 @@
+#include "hittable_list.h"
+#include "ray.h"
+
+bool hittable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const
+{
+  hit_record temp_rec;
+  bool hit_anything = false;
+
+  auto closest_so_far = t_max;
+
+  for(const auto& object : objects)
+    {
+      if(object->hit(r, t_min, closest_so_far, temp_rec))
+        {
+          hit_anything = true;
+          closest_so_far = temp_rec.t;
+          rec = temp_rec;
+        }
+    }
+  return hit_anything;
+}
