+++ src/common/hittable_list.h
@@ -22,24 +22,4 @@ class hittable_list : public hittable
 public:
   std::vector<unique_ptr<hittable>> objects;
 };
-
-bool hittable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const
-{
-  hit_record temp_rec;
-  bool hit_anything = false;
-
-  auto closest_so_far = t_max;
-
-  for(const auto& object : objects)
-    {
-      if(object->hit(r, t_min, closest_so_far, temp_rec))
-        {
-          hit_anything = true;
-          closest_so_far = temp_rec.t;
-          rec = temp_rec;
-        }
-    }
-  return hit_anything;
-}
-
 #endif // HITTABLE_LIST_HH_INCLUDED
