+++ src/simple_tracer/marble_single_thread.cc
@@ -8,6 +8,7 @@
 #include "camera.h"
 #include "material.h"
 #include <cassert>
+#include "sphere_generation.h"
 
 hittable_list static_scene()
 {
@@ -23,62 +24,6 @@ hittable_list static_scene()
   world.add(make_unique<sphere>(point3(0, -1000, 0), 1000, std::move(ground_material)));
   world.add(make_unique<sphere>(point3(5,1.5,1),1,std::move(glass_material)));
   world.add(make_unique<sphere>(point3(1,1,1),1,std::move(metal_material)));
-  return world;
-}
-
-hittable_list random_scene()
-{
-  hittable_list world;
-
-  auto ground_material = make_unique<lambertian>(color(0.5, 0.5, 0.5));
-  world.add(make_unique<sphere>(point3(0, -1000, 0), 1000, std::move(ground_material)));
-
-  for (int a = -11; a < 11; a++)
-    {
-      for(int b = -11; b < 11; b++)
-        {
-          auto choose_mat = random_double();
-          point3 center(a+0.9*random_double(), 0.2, b + 0.9*random_double());
-
-          if((center - point3(4, 0.2, 0)).length() > 0.9)
-            {
-              unique_ptr<material> sphere_material;
-
-              if(choose_mat < 0.8)
-                {
-                  // diffuse
-                  auto albedo = color::random() * color::random();
-                  sphere_material = make_unique<lambertian>(albedo);
-                  world.add(make_unique<sphere>(center, 0.2, std::move(sphere_material)));
-
-                }
-              else if(choose_mat < 0.95)
-                {
-                  // metal
-                  auto albedo = color::random(0.5, 1);
-                  auto fuzz = random_double(0, 0.5);
-                  sphere_material = make_unique<metal>(albedo, fuzz);
-                  world.add(make_unique<sphere>(center, 0.2, std::move(sphere_material)));
-                }
-              else
-                {
-                  // glass
-                  sphere_material = make_unique<dielectric>(1.5);
-                  world.add(make_unique<sphere>(center, 0.2, std::move(sphere_material)));
-
-                }
-            }
-        }
-  }
-  auto material1 = make_unique<dielectric>(1.5);
-  world.add(make_unique<sphere>(point3(0, 1, 0), 1.0, std::move(material1)));
-
-  auto material2 = make_unique<lambertian>(color(0.4, 0.2, 0.1));
-  world.add(make_unique<sphere>(point3(-4, 1, 0), 1.0, std::move(material2)));
-
-  auto material3 = make_unique<metal>(color(0.7, 0.6, 0.5), 0.0);
-  world.add(make_unique<sphere>(point3(4, 1, 0), 1.0, std::move(material3)));
-
   return world;
 }
 
@@ -114,8 +59,9 @@ int main()
   const int samples_per_pixel = 500;
   const int max_depth = 10;
 
+  SphereGeneration scene_generator;
   // World
-  auto world = random_scene();
+  auto world = scene_generator.random_scene_hittablelist(11);
 
   point3 lookfrom(13, 2, 3);
   point3 lookat(0, 0, 0);
